+++ HISTORY.rst
@@ -6,12 +6,13 @@ History
 1.1 (2019-06-??)
 ______________________
 * Use the PartInfo KitSpace API.
-* Creates a KiCad plugin integration and desktop shortcuts (not im Mac-OS) during the installation.
+* Creates a KiCad plugin integration and desktop shortcuts (not in Mac-OS) during the installation.
 * Add KiCost to the context menu "Open with..." in Windows to XML and CSV files.
-* Show (and allow to change) the current currency rate for distributors that are not in the specified curreny.
+* Show (and allow to change in the spreadsheet) the current currency rate for distributors that are not in the specified curreny.
 * Highlight non-active in production / not-recommended-to-new-layout / decrepit components.
 * Improved the spreadsheet purchase distributors code formula and fix the Digikey/Mouser import errors.
 * Show the total purchased at the spreadsheet.
+* Split the GUI file in programming one in wxFormBuilder generated.
 * Fix the erro when deal with scapes in the names of BOM files on `test.sh`.